requirements.txt
requests==2.25.0
Flask_WTF==0.14.3
WTForms==2.3.3
Flask==1.1.2
beautifulsoup4==4.9.3
mysql_connector_repackaged==0.3.1
